+++ core/data/src/main/java/soft/divan/financemanager/core/data/dto/CreateAccountRequestDto.kt
@@ -1,6 +1,7 @@
 package soft.divan.financemanager.core.data.dto
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class CreateAccountRequestDto(
     @SerializedName("id")
@@ -8,7 +9,7 @@ data class CreateAccountRequestDto(
     @SerializedName("name")
     val name: String,
     @SerializedName("balance")
-    val balance: Double,
+    val balance: BigDecimal,
     @SerializedName("currencyId")
     val currencyId: String
 )
